releasetools: Regenerate hashtree for prebuilt odm.img files
When building with a prebuilt odm.img file (via BOARD_PREBUILT_ODMIMAGE)
the odm.img may not have the same hashtree and salt as the ones being
created during the build process.

To make sure that odm.img matches system/boot/dtbo, the hashtree should be
rewritten using avbtool.

This patch therefor adds in the rewrite of the hashtree if using a prebuilt
odm.img and AVB is enabled.  The code is similar to the code in the
Add_Dtbo() function in this same script.
releasetools: Regenerate hashtree for prebuilt odm.img files

Change-Id: I365491aabbe51e4757bfc9f1cf317e07306300a6